requirements.txt
websocket-client
requests
httpx
numpy
pandas
python-dotenv